windowsssss
Former-commit-id: d564c7f57e6aadc9fc2bd7882353bace5c74ad39 [formerly 307a10bf11a01cda0c5822ad1126483c9d46e822]
Former-commit-id: e5416cc187da27031bf6e037b2c3a2bfafd5c3c6

+++ Core/Window.cpp
@@ -33,7 +33,7 @@ void Window::CursorPosCallback(GLFWwindow* window, double x, double y) {
 	win->mInput->mCurrent.mCursorPos.y = (float)y + win->ClientRect().offset.y;
 
 	if (win->mTargetCamera) {
-		float2 clip = 2.f * (float2(x,y) / float2(win->ClientRect().extent.width, win->ClientRect().extent.height)) - 1.0f;
+		float2 clip = 2.f * (float2((float)x, (float)y) / float2((float)win->ClientRect().extent.width, (float)win->ClientRect().extent.height)) - 1.0f;
 		win->mInput->mMousePointer.mWorldRay.mOrigin = win->mTargetCamera->WorldPosition();
 		win->mInput->mMousePointer.mWorldRay.mDirection = win->mTargetCamera->ClipToWorldRay(float3(clip, 0.f));
 	}
